Define rebind template struct in fftw_allocator<T> to compile with GCC 9.
Newer versions of the STL library apparently use rebind in places they did not
before. Because fftw_allocator<T> inherits from std::allocator<T>, it also
inherits that rebind unless it is overridden. This caused std::vector to be
unable to compile with fftw_allocator due to the mismatch of allocator types.